+++ Tests/SuperwallKitTests/Paywall/Presentation/Internal Presentation/Operators/StorePresentationObjectsOperatorTests.swift
@@ -20,7 +20,7 @@ final class StorePresentationObjectsOperatorTests: XCTestCase {
     let input = PresentablePipelineOutput(
       request: request,
       debugInfo: [:],
-      paywallViewController: dependencyContainer.makePaywallViewController(for: .stub(), withCache: nil),
+      paywallViewController: dependencyContainer.makePaywallViewController(for: .stub(), withCache: nil, delegate: nil),
       presenter: UIViewController(),
       confirmableAssignment: nil
     )